+++ components/src/lib/index.ts
@@ -1,13 +1,13 @@
-import ModalContainer from './ModalContainer';
-import ModalContent from './ModalContent';
-import ModalBody from './ModalBody';
-import ModalCloseButton from './ModalCloseButton';
-import ModalOverlay from './ModalOverlay';
-import ModalTitle from './ModalTitle';
-import CancelButton from './CancelButton';
-import ConfirmButton from './ConfirmButton';
-import AlertActions from './AlertActions';
-import Input from './Input';
+import ModalContainer from './components/Modal/ModalContainer';
+import ModalContent from './components/Modal/ModalContent';
+import ModalBody from './components/Modal/ModalBody';
+import ModalCloseButton from './components/Modal/ModalCloseButton';
+import ModalOverlay from './components/Modal/ModalOverlay';
+import ModalTitle from './components/Modal/ModalTitle';
+import CancelButton from './components/Modal/CancelButton';
+import ConfirmButton from './components/Modal/ConfirmButton';
+import AlertActions from './components/Modal/AlertActions';
+import Input from './components/Modal/Input';
 
 export const Modal = Object.assign(ModalContainer, {
   Overlay: ModalOverlay,
